+++ test_interactions.py
@@ -210,12 +210,12 @@ async def test_yield():
 
     r=await hr.init()
 
-    assert o._contents==[]
+    assert o.childs==[]
 
     r=await hr.interact(o).inc()
 
     # first yield, yield nothing
-    assert o._contents==[]
+    assert o.childs==[]
 
     assert r["next"]
     assert "stream" not in r
@@ -223,7 +223,7 @@ async def test_yield():
     r=await hr.doNext( r)
 
     # seconf yield, yield list("abc")
-    assert o._contents==['a', 'b', 'c']
+    assert o.childs==['a', 'b', 'c']
     assert "stream" in r
     assert r["stream"][id(o)]=="abc"
     assert r["next"]
@@ -231,7 +231,7 @@ async def test_yield():
 
     r=await hr.doNext( r)
 
-    assert o._contents==['a', 'b', 'c', 'd']
+    assert o.childs==['a', 'b', 'c', 'd']
     assert "stream" in r
     assert r["stream"][id(o)]=="d"
     assert r["next"]